refactor(app): put auth functions together
